refactor explorer table folder to sub folders, separate constants from types
