+++ packages/client/src/pages/Query/Explorer/ExplorerTable/Cell/ExplorerCellOverflowStyles.tsx
