+++ packages/client/src/pages/Query/Explorer/ExplorerTable/ExplorerTableRow.tsx
@@ -19,7 +19,7 @@ import { EntityEnums, RelationEnums } from "@inkvisitor/shared/enums";
 import { UserTagSize } from "components/advanced/UserTag/utils";
 import { invalidateAllExplorerQueries } from "pages/Query/useQueryData";
 import { getRelationSuggesterConfig } from "pages/Query/utils";
-import { CELL_DISPLAY_LIMIT, ExplorerCellOverflow } from "./ExplorerCellOverflow";
+import { CELL_DISPLAY_LIMIT, ExplorerCellOverflow } from "./Cell/ExplorerCellOverflow";
 import {
   StyledAltLabelAddInput,
   StyledAltLabelChip,
@@ -35,7 +35,7 @@ import {
   StyledFocusedCircle,
   StyledRowInner,
 } from "./ExplorerTableStyles";
-import { WIDTH_COLUMN_DEFAULT, WIDTH_COLUMN_EUC, WIDTH_COLUMN_FIRST } from "./types";
+import { WIDTH_COLUMN_DEFAULT, WIDTH_COLUMN_EUC, WIDTH_COLUMN_FIRST } from "./constants";
 
 const EditableCellValue: React.FC<{
   value: string;
