+++ packages/client/src/pages/Query/Explorer/ExplorerTable/Header/ExploreTableHeaderTooltip.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { IEntity } from "@inkvisitor/shared/types";
 import { Explore } from "@inkvisitor/shared/types/query";
 import { Tooltip } from "components";
-import { renderExploreColumnParamValue } from "./ExploreColumnParamValueRenderers";
+import { renderExploreColumnParamValue } from "../ExplorerTableNewColumnPanel/ExploreColumnParamValueRenderers";
 import {
   StyledTooltipRow,
   StyledTooltipLabel,
