+++ packages/client/src/pages/Query/Explorer/ExplorerTable/constants.ts
@@ -1,0 +1,4 @@
+export const WIDTH_COLUMN_FIRST = 280;
+export const WIDTH_COLUMN_DEFAULT = 400;
+export const WIDTH_COLUMN_EUC = 210;
+export const HEIGHT_ROW_DEFAULT = 38;
