+++ packages/client/src/pages/Query/Explorer/ExplorerTable/types.ts
@@ -29,7 +29,3 @@ export const restrictedBatchActions = new Set<BatchAction>([
   BatchAction.add_relation,
 ]);
 
-export const WIDTH_COLUMN_FIRST = 280;
-export const WIDTH_COLUMN_DEFAULT = 400;
-export const WIDTH_COLUMN_EUC = 210;
-export const HEIGHT_ROW_DEFAULT = 38;
