+++ packages/client/src/pages/Query/constants.ts
@@ -1,0 +1,10 @@
+import { Query } from "@inkvisitor/shared/types";
+
+export const QUERY_GRID_WIDTH = 400;
+export const QUERY_GRID_HEIGHT = 50;
+
+export const QUERY_LEFT_PANEL_MIN_WIDTH = 410;
+export const QUERY_RIGHT_PANEL_MIN_WIDTH = 450;
+export const QUERY_PAGE_SEPARATOR_X_PERCENT_POSITION = 60;
+export const QUERY_SEARCH_PANEL_MIN_HEIGHT = 34;
+export const QUERY_BUILDER_MIN_HEIGHT = 150;
